pyproject.toml
[project]
name = "echo-backend"
version = "0.1.0"
description = "Echo MVP backend scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "sqlalchemy[asyncio]>=2.0.32",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "python-decouple>=3.8",
  "boto3>=1.35.0",
  "apscheduler>=3.10.4",
  "supabase>=2.7.4",
  "python-multipart>=0.0.9",
  "hume>=0.7.0",
  "openai>=1.51.0",
  "stripe>=10.8.0",
  "redis>=5.0.8",
  "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "aiosqlite>=0.20.0",
  "pytest-asyncio>=0.23.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
